kernel/thor: Move trap handler to its own file
