+++ kernel/thor/arch/riscv/thor-internal/arch/trap.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+extern "C" void thorExceptionEntry();
